Update Arduino Test Programs/Receive_setup_from_GUI/Receive_setup_from_GUI.ino
Fixed serial reading during check status
The driver tx would hang on disconnect, likely because the port it was writing too was no longer available.  The problem was that the Arduino would need over a second, after the port was closed to parse the incoming packet and respond.  As a ll packets are critical, the circular buffer was replaced with a full read of the serial buffer on a serial event, allowing the Arduino to fully parse the disconnect command when the GUI and com port are closed.